fix: make it work with latest version in github branch
